add simple example
