+++ website/src/main.js
@@ -1,9 +1,12 @@
-import Vue from 'vue'
+import Vue from 'vue';
+import router from './router';
 import App from './App.vue'
-import vuetify from './plugins/vuetify'
+import vuetify from './plugins/vuetify';
+
 Vue.config.productionTip = false
 
 new Vue({
+  router,
   vuetify,
   render: h => h(App)
 }).$mount('#app')
