+++ website/src/plugins/vuetify.js
@@ -4,4 +4,12 @@ import Vuetify from 'vuetify/lib/framework';
 Vue.use(Vuetify);
 
 export default new Vuetify({
+    themes: {
+        light: {
+          primary: '#3f51b5',
+          secondary: '#b0bec5',
+          anchor: '#8c9eff',
+        },
+        dark: false,
+    },
 });
