+++ website/src/views/HomePage.vue
@@ -1,0 +1,22 @@
+<template>
+  <v-row style="margin-top: 20px">
+    <v-card class="mx-auto" width="400">
+      <v-list subheader three-line>
+        <v-list-item>
+          <v-list-item-content>
+            <v-list-item-title>
+              <a href="/simple">Simple Example</a>
+            </v-list-item-title>
+          </v-list-item-content>
+        </v-list-item>
+        <v-list-item>
+          <v-list-item-content>
+            <v-list-item-title>
+              <a href="/cart-checkout-example">Cart Checkout Example</a>
+            </v-list-item-title>
+          </v-list-item-content>
+        </v-list-item>
+      </v-list>
+    </v-card>
+  </v-row>
+</template>
